Make OCI spec opts available on all platforms
